requirements.txt
aniso8601==9.0.1
apispec==6.3.0
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachelib==0.9.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Deprecated==1.2.14
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Caching==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
limits==3.5.0
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.19.0
mdurl==0.1.2
numpy==1.24.3
ordered-set==4.1.0
osmium==3.6.0
packaging==23.1
pandas==2.0.1
Pygments==2.15.1
PyJWT==2.7.0
pymongo==4.3.3
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
ratelimit==2.2.1
requests==2.30.0
rich==13.4.2
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
typing_extensions==4.6.0
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.3.4
wrapt==1.15.0
